refactor: move agent runtime host leaves to daemon
