+++ packages/daemon/src/agent-runtime/host-discovery.ts
